before deploy practice

+++ src/controllers/videoControllers.js
@@ -1,14 +1,17 @@
 import Video from "../models/Video";
 import User from "../models/User";
+import Comment from "../models/Comment";
 
 export const home = async (req, res) => {
-  const videos = await Video.find({}).sort({ creationTime: "desc" });
-  res.render("video/home", { pageTitle: "HOME", videos });
+  const videos = await Video.find({})
+    .sort({ creationTime: "desc" })
+    .populate("owner");
+  return res.render("video/home", { pageTitle: "HOME", videos });
 };
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id).populate("owner");
+  const video = await Video.findById(id).populate("owner").populate("comments");
   if (!video) {
     return res.render("template/404", { pageTitle: "404" });
   } else {
@@ -18,6 +21,11 @@ export const watch = async (req, res) => {
 
 export const getUpload = async (req, res) => {
   return res.render("video/upload", { pageTitle: "Upload" });
+};
+
+export const thumbnail = (req, res, next) => {
+  const { id } = req.body;
+  const newVideo = Video.findById(id);
 };
 
 export const postUpload = async (req, res) => {
@@ -40,7 +48,6 @@ export const postUpload = async (req, res) => {
     const user = await User.findById(_id);
     user.videos.push(newVideo._id);
     user.save();
-
     return res.redirect("/");
   } catch (err) {
     return res.status(400).render("video/upload", {
@@ -60,6 +67,7 @@ export const getEdit = async (req, res) => {
       .render("template/404", { pageTitle: "404 NOT FOUND" });
   }
   if (String(video.owner) !== String(_id)) {
+    req.flash("error", "권한이 없습니다");
     return res.status(403).redirect("/");
   }
   return res.render("video/edit", {
@@ -83,6 +91,7 @@ export const postEdit = async (req, res) => {
       .render("template/404", { pageTitle: "404 NOT FOUND" });
   }
   if (String(video.owner) !== String(_id)) {
+    req.flash("error", "권한이 없습니다");
     return res.status(403).rendirect("/");
   }
   await Video.findByIdAndUpdate(id, {
@@ -96,7 +105,17 @@ export const postEdit = async (req, res) => {
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
   const { _id } = req.session.user;
-  const video = await Video.findByIdAndDelete(id);
+  const video = await Video.findById(id);
+  if (String(video.owner) !== String(_id)) {
+    req.flash("error", "영상의 게시자만 영상을 삭제할 수 있습니다");
+    return res.status(403).redirect(`/videos/${id}`);
+  }
+  await Video.findByIdAndDelete(id);
+  const user = await User.findById(_id);
+  const newVideos = user.videos.filter(
+    (element) => String(element._id) !== String(id)
+  );
+  user.videos = newVideos;
   return res.redirect("/");
 };
 
@@ -108,10 +127,69 @@ export const search = async (req, res) => {
       title: {
         $regex: new RegExp(`${keyword}`, "i"),
       },
-    });
+    }).populate("owner");
   }
-  if (String(video.owner) !== _id) {
-    return res.status(403).rendirect("/");
-  }
+
   return res.render("video/search", { pageTitle: "Search", videos });
+};
+
+export const registerView = async (req, res) => {
+  const { id } = req.params;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.sendStatus(404);
+  }
+  video.meta.views = video.meta.views + 1;
+  await video.save();
+  return res.sendStatus(200);
+};
+
+export const createComment = async (req, res) => {
+  const {
+    params: { id },
+    session: { user },
+    body: { text },
+  } = req;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.sendStatus(404);
+  }
+  const comment = await Comment.create({
+    text,
+    owner: user._id,
+    video: id,
+  });
+  video.comments.push(comment._id);
+  video.save();
+  const commentWriterId = comment.owner._id;
+  const commentWriter = await User.findById(commentWriterId);
+  commentWriter.comments.push(comment._id);
+  commentWriter.save();
+  return res.status(201).json({ newCommentId: comment._id });
+};
+
+export const deleteComment = async (req, res) => {
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+      user,
+    },
+  } = req;
+  const comment = await Comment.findById(id).populate("owner");
+  if (!comment) {
+    return res
+      .status(404)
+      .render("template/404", { pageTitle: "404 NOT FOUND" });
+  }
+  const targetUser = await User.findById(_id);
+  if (String(comment.owner._id) === String(_id)) {
+    await Comment.findByIdAndDelete(id);
+    const newComments = targetUser.comments.filter(
+      (element) => String(comment.id) !== String(element._id)
+    );
+    targetUser.comments = newComments;
+    targetUser.save();
+    return res.sendStatus(204);
+  }
 };
